Fix documentation warnings about resolving Path
Due to limitations in Sphinx, an abstract parent dataclass with
type Path creates problems generating documentation for child
classes defined in other modules that don't import Path. For
dependencies, since these are pure models, move them into a single
neophile.models.dependencies module, which resolves the problem.
For updates, since they carry business logic, for now import Path
into each module with a noqa annotation.
