--- src/neophile/dependency/kustomize.py
@@ -1,26 +1,0 @@
-"""A Kustomize external dependency."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass
-
-from neophile.dependency.base import Dependency
-
-__all__ = ["KustomizeDependency"]
-
-
-@dataclass(frozen=True, order=True)
-class KustomizeDependency(Dependency):
-    """Represents a single Kustomize dependency."""
-
-    url: str
-    """The full URL of the dependency."""
-
-    owner: str
-    """The owner of the referenced GitHub repository."""
-
-    repo: str
-    """The name of the referenced GitHub repository."""
-
-    version: str
-    """The version of the dependency."""
