+++ src/neophile/update/kustomize.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import re
 from dataclasses import dataclass
+from pathlib import Path  # noqa: F401
 
 from ruamel.yaml import YAML
